requirements.txt
python-socketio==5.11.2
PySide6==6.9.2
msgpack==1.0.8
requests>=2.31.0
websocket-client>=1.7.0
tzlocal>=5.0
numpy>=1.26.0
pandas>=2.2.0
firebase-admin>=6.6.0
google-cloud-firestore>=2.20.0
google-cloud-storage>=3.1.0
supabase>=2.0.0
python-dotenv>=1.0.0